python-logic/requirements.txt
ultralytics==8.0.26
numpy
matplotlib
lapx
pymongo
gridfs
pillow
asyncio
python-opencv
telegram
